New Particle system
Added particles and improved removing them off the screen
Added sound effects (requires user interaction with the document [e.g. 'onclick'])
Now game objects with sprites can only be rendered after the image is fully loaded, this only works if the current game object in the loop (for loop / for each loop) has a prop. named type and if that type is "GameObject" or "Sprite" is going to check if the 'isLoaded' prop. (on the GameObject / Sprite class) is true. 
All the other objects without a type get rendered as normal (else statement)